added pos and neg to array time series
