+++ timeseries/arraytimeseries.py
@@ -37,7 +37,7 @@ class ArrayTimeSeries(SizedContainerTimeSeriesInterface):
 
         if len(self._times) != len(set(self._times)):
             raise ValueError("Time data should contain no repeats")
-    
+
     def __len__(self):
         # Note: len of np.appry returns and error for arrays of size 1.
         #super().__len__()
@@ -48,6 +48,8 @@ class ArrayTimeSeries(SizedContainerTimeSeriesInterface):
 
     def __eq__(self, rhs):
         # Note: np.array_equal compares both elements and dimensions.
+        #print("array check",self,rhs)
+        #print("type check",type(self),type(rhs))
         self.__class__._check_time_domains_helper(self, rhs)
         try:
             return (np.array_equal(self._values,rhs._values) and np.array_equal(self._times,rhs._times))
@@ -55,56 +57,6 @@ class ArrayTimeSeries(SizedContainerTimeSeriesInterface):
         except TypeError:
             raise NotImplemented
 
-    # imo we should consider implementing this in
-    # SizedContainerTimeSeriesInterface...
-    def interpolate(self,ts_to_interpolate):
-        """
-        Returns new TimeSeries instance with piecewise-linear-interpolated values
-        for submitted time-times.If called times are outside of the domain of the existing
-        Time Series, the minimum or maximum values are returned.
-
-        Parameters
-        ----------
-        self: TimeSeries instance
-        ts_to_interpolate: list or other sequence of times to be interpolated
-
-        """
-        def binary_search(times, t):
-            """ Returns surrounding time indexes for value that is to be interpolated"""
-            min = 0
-            max = len(times) - 1
-            while True:
-                if max < min:
-                    return (max,min)
-                m = (min + max) // 2
-                if times[m] < t:
-                    min = m + 1
-                elif times[m] > t:
-                    max = m - 1
-                else: #Should never hit this case in current implementation
-                    return (min,max)
-
-        def interpolate_val(times,values,t):
-            """Returns interpolated value for given time"""
-
-            if t in times:          #time already exits in ts -- return it
-                return values[times.index(t)]
-
-            elif t >= times[-1]:    #time is above the domain of the existing values -- return max time value
-                return values[-1]
-
-            elif t <= times[0]:     #time is below the domain of the existing values -- return min time value
-                return values[0]
-
-            else:                   #time is between two existing points -- interpolate it
-                low,high = binary_search(times, t)
-                slope = (float(values[high]) - values[low])/(times[high] - times[low])
-                c = values[low]
-                interpolated_val = (t-times[low])*slope + c
-                return interpolated_val
-
-        interpolated_ts = [interpolate_val(self._times,self._values,t) for t in ts_to_interpolate]
-        return self.__class__(values=interpolated_ts,times=ts_to_interpolate)
 
     # J: these methods need to be implemented
     def __add__ (self):
@@ -114,9 +66,11 @@ class ArrayTimeSeries(SizedContainerTimeSeriesInterface):
     def __ne__(self):
         pass
     def __neg__(self):
-        pass
+        return self.__class__(values=((-1)*self._values), times=self._times)
+
     def __pos__(self):
-        pass
+        return self.__class__(values=self._values, times=self._times)
+
     def __radd__(self):
         pass
     def __rmul__(self):
