+++ timeseries/sizedcontainertimeseriesinterface.py
@@ -285,41 +285,6 @@ class SizedContainerTimeSeriesInterface(TimeSeriesInterface):
     def items(self):
         return list(zip(self._times, self._values))
 
-    def __neg__(self):
-        return self.__class__((-x for x in self._values), self._times)
-
-    def __pos__(self):
-        return self.__class__((x for x in self._values), self._times)
-
-    def __add__(self, rhs):
-        try:
-            if isinstance(rhs, numbers.Real):
-                # R: may be worth testing time domains are preserved correctly
-                return self.__class__((a + rhs for a in self), self._times)
-            else:
-                self._check_length_helper(self, rhs)
-                # R: test me. should fail when the time domains are non congruent
-                self._check_time_domains_helper(self, rhs)
-                pairs = zip(self._values, rhs)
-                return self.__class__((a + b for a, b in pairs), self._times)
-        except TypeError:
-            raise NotImplemented # R: test me. should fail when we try to add a numpy array or list
-
-
-    def __radd__(self, other): # other + self delegates to self.__add__
-        return self + other
-
-    def __sub__(self, rhs):
-        try:
-            if isinstance(rhs, numbers.Real):
-                return self.__class__((a - rhs for a in self), self._times)
-            else:
-                self._check_length_helper(self, rhs)
-                self._check_time_domains_helper(self, rhs)
-                pairs = zip(self._values, rhs)
-                return self.__class__((a - b for a, b in pairs), self._times)
-        except TypeError:
-            raise NotImplemented
 
     def interpolate(self,ts_to_interpolate):
         """
