+++ timeseries/timeseries.py
@@ -100,6 +100,43 @@ class TimeSeries(SizedContainerTimeSeriesInterface):
 
 
     # J: should this not be (self - other)?
+
+    def __neg__(self):
+        return self.__class__((-x for x in self._values), self._times)
+
+    def __pos__(self):
+        return self.__class__((x for x in self._values), self._times)
+
+    def __add__(self, rhs):
+        try:
+            if isinstance(rhs, numbers.Real):
+                # R: may be worth testing time domains are preserved correctly
+                return self.__class__((a + rhs for a in self), self._times)
+            else:
+                self._check_length_helper(self, rhs)
+                # R: test me. should fail when the time domains are non congruent
+                self._check_time_domains_helper(self, rhs)
+                pairs = zip(self._values, rhs)
+                return self.__class__((a + b for a, b in pairs), self._times)
+        except TypeError:
+            raise NotImplemented # R: test me. should fail when we try to add a numpy array or list
+
+
+    def __radd__(self, other): # other + self delegates to self.__add__
+        return self + other
+
+    def __sub__(self, rhs):
+        try:
+            if isinstance(rhs, numbers.Real):
+                return self.__class__((a - rhs for a in self), self._times)
+            else:
+                self._check_length_helper(self, rhs)
+                self._check_time_domains_helper(self, rhs)
+                pairs = zip(self._values, rhs)
+                return self.__class__((a - b for a, b in pairs), self._times)
+        except TypeError:
+            raise NotImplemented
+
     def __rsub__(self, other):
         return -(self - other)
